Ya terminado todo lo que se refiere a las operaciones y eso. Faltan los test eso.

+++ src/postfixcalc/IPostfixCalc.java
@@ -1,5 +1,5 @@
 package postfixcalc;
 
 public interface IPostfixCalc {
-	double Evaluate(int val1, int val2, String operador);
+	int Evaluate(int val1, int val2, String operador);
 }
